generalize the retry delay callable that can also be used for something else like logging
the callback will also receive the exception as parameter because that is the only information that would otherwise not be accessible. with this approach it is also possible to have a callback that does decides if retrying makes sense based on the exception message. so when the exception message does not match, the callback can simply rethrow the exception which will also stop the retry logic